Moved redistribution components to a new redistribution folder (#389)
