+++ Assets/Scripts/Algorithms/Patrolling/PartitionedAlgorithms/PartitionedConscientiousReactive.cs
@@ -25,6 +25,7 @@
 using System.Linq;
 
 using Maes.Algorithms.Patrolling.Components;
+using Maes.Algorithms.Patrolling.Components.Redistribution;
 using Maes.Map;
 using Maes.Robot;
